+++ library/QATools/QATools/BEM/BEMPageFactory.php
@@ -15,6 +15,7 @@ use QATools\QATools\BEM\ElementLocator\BEMElementLocatorFactory;
 use QATools\QATools\BEM\ElementLocator\LocatorHelper;
 use QATools\QATools\BEM\PropertyDecorator\BEMPropertyDecorator;
 use QATools\QATools\PageObject\Config\IConfig;
+use QATools\QATools\PageObject\Container;
 use QATools\QATools\PageObject\ISearchContext;
 use QATools\QATools\PageObject\PageFactory;
 use QATools\QATools\PageObject\PropertyDecorator\IPropertyDecorator;
@@ -36,15 +37,15 @@ class BEMPageFactory extends PageFactory
 	/**
 	 * Creates BEMPageFactory instance.
 	 *
-	 * @param Session       $session        Mink session.
-	 * @param IConfig       $config         Page factory configuration.
-	 * @param LocatorHelper $locator_helper Locator helper.
+	 * @param Session        $session        Mink session.
+	 * @param Container|null $container      Dependency injection container.
+	 * @param LocatorHelper  $locator_helper Locator helper.
 	 */
-	public function __construct(Session $session, IConfig $config = null, LocatorHelper $locator_helper = null)
+	public function __construct(Session $session, Container $container = null, LocatorHelper $locator_helper = null)
 	{
 		$this->annotationRegistry['bem'] = '\\QATools\\QATools\\BEM\\Annotation\\BEMAnnotation';
 
-		parent::__construct($session, $config);
+		parent::__construct($session, $container);
 
 		$this->_locatorHelper = isset($locator_helper) ? $locator_helper : new LocatorHelper();
 	}