+++ CHANGELOG.md
@@ -11,6 +11,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - When "TypifiedElementProxy" class was used manually (not through annotations) the element class was "TextBlock" instead of "AbstractTypifiedElement".
 - When "TypifiedElementCollectionProxy" class was used manually (not through annotations) the element class was "TextBlock" instead of "AbstractTypifiedElementCollection".
 - All elements created manually (not through annotations) now require `IPageFactory` instance as 2nd argument (before only container type elements were needing this).
+- The second optional parameter of `PageFactory` classes is now dependency injection container, instead of a `Config`.
+- Following methods of `PageFactory` class are not longer part of public API: `setAnnotationManager`, `setSession`.
+
+### Removed
+- Following methods were removed from `PageFactory` class: `getAnnotationManager`, `setUrlFactory`, `getUrlFactory`, `setUrlNormalizer`, `setPageLocator`.
 
 ### Fixed
 - When "WebElementCollectionProxy" class was used manually (not through annotations) the element class was "WebElement" instead of "WebElementCollection".